fix(ui): preserve agent filters in generated search code
