+++ ui-next/src/pages/agent/executions/workflowSearchComponents/AdvancedSearch.tsx
@@ -365,6 +365,8 @@ export default function AdvancedSearch({
                 sort,
                 freeText,
                 query: buildQuery().query,
+                classifier: "agent",
+                topLevelOnly: hideSubWorkflows,
               }}
             />
           )}
